refactor: JwtService 인터페이스를 추가한다
